interval: update loopStruct to locate original head and latch
This is a rather ugly commit, and it seem to diverge from
the line-by-line description of the Structuring Algorithm
described in Cifuentes' Reverse Compilation Techniques.

However, without this change. We can identify loops in derived
sequences of graphs well, but we can't relate them back to
thier original nodes. And as such, the information we get is
not quite useful for later stages of control flow recovery.